tests/requirements.txt
# Generated by make generate_requirements
astroid==2.3.3
attrs==19.3.0
behave==1.2.6
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.9
flake8-commas==2.0.0
flake8-print==3.1.4
flake8-quotes==2.1.1
flake8-string-format==0.3.0
humanfriendly==7.1.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
Jinja2==2.11.1
jsonschema==3.2.0
kazoo==2.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.1
parse==1.14.0
parse-type==0.5.2
pep8==1.7.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyHamcrest==2.0.0
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
PyYAML==5.3
requests==2.23.0
six==1.14.0
tenacity==6.0.0
texttable==1.6.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
websocket-client==0.57.0
wrapt==1.11.2
yapf==0.29.0
zipp==3.0.0
